chore: scale scripts folder structure to support multichain deploy scripts
