+++ scripts/generic/create-protection-market-curve3Crypto.ts
@@ -1,14 +1,27 @@
 /**
- * @notice Deploys the `CurveThreeTokens` trigger and creates a protection market
- * @dev Ensure the desired TEST_CHAIN_FORK and RPC_URL env variables are set to target a specific chain
- * @dev To deploy, you must define the CURVE_POOL environment variable, which is the address of the curve pool on the desired chain
+ * @notice Deploys the `CurveThreeTokenBasePool` trigger and creates a protection market
+ * @dev You must define the CURVE_POOL and MULTISIG environment variables, which are the addresses of the Curve pool and Cozy multisig on the desired chain.
+ * Sample deploy commands are below:
+ *
+ *     Deploy a mainnet Curve tricrypto pool protection market:
+ *         CURVE_POOL=0xD51a44d3FaE010294C616388b506AcdA1bfAAE46 MULTISIG=0x1725d89c5cf12F1E9423Dc21FdadC81C491a868b yarn hardhat run scripts/create-protection-market-curve3Crypto.ts --network mainnet
+ *
+ *     Deploy an Arbitrum Curve tricrypto pool protection market:
+ *         CURVE_POOL=0x960ea3e3C7FB317332d990873d354E18d7645590 MULTISIG=0xe570e347932621d08c242035c56d79a33f4269af yarn hardhat run scripts/create-protection-market-curve3Crypto.ts --network arbitrum
+ *
+ *     Deploy a mainnet Curve tricrypto pool protection market, on an instance of Hardhat that forks mainnet:
+ *         CURVE_POOL=0xD51a44d3FaE010294C616388b506AcdA1bfAAE46 MULTISIG=0x1725d89c5cf12F1E9423Dc21FdadC81C491a868b yarn hardhat run scripts/create-protection-market-curve3Crypto.ts
+ *
+ *     Deploy an Abitrum Curve tricrypto pool protection market, on an instance of Hardhat that forks Arbitrum:
+ *         TEST_CHAIN_FORK=arbitrum CURVE_POOL=0x960ea3e3C7FB317332d990873d354E18d7645590 MULTISIG=0xe570e347932621d08c242035c56d79a33f4269af yarn hardhat run scripts/create-protection-market-curve3Crypto.ts
+ *
  */
 import hre from 'hardhat';
 import '@nomiclabs/hardhat-ethers';
 import { Contract, ContractFactory, utils } from 'ethers';
 import chalk from 'chalk';
-import { getChainId, getContractAddress, getGasPrice, logSuccess, logFailure, findLog, waitForInput, fundAccount } from '../utils/utils'; // prettier-ignore
-import comptrollerAbi from '../abi/Comptroller.json';
+import { getChainId, getContractAddress, getGasPrice, logSuccess, logFailure, findLog, waitForInput } from '../../utils/utils'; // prettier-ignore
+import comptrollerAbi from '../../abi/Comptroller.json';
 
 // STEP 0: ENVIRONMENT SETUP
 const provider = hre.ethers.provider;
@@ -31,20 +44,28 @@ async function main(): Promise<void> {
   const curvePoolAddress = String(process.env.CURVE_POOL);
   if (!curvePoolAddress) throw new Error("Please define the 'CURVE_POOL' environment variable.");
 
+  const cozyMultiSig = String(process.env.MULTISIG);
+  if (!cozyMultiSig) throw new Error("Please define the 'MULTISIG' environment variable.");
+
   if (!utils.isAddress(recipient)) throw new Error('\n\n**** Please set the recipient address on line 23 ****\n');
 
   // Compile contracts to make sure we're using the latest version of the trigger contracts
   await hre.run('compile');
 
-  const overrides = { gasPrice: await getGasPrice() };
+  // Do some preparation
+  const underlyingAddress = getContractAddress('ETH', chainId);
+  const overrides = await getGasPrice();
 
   // VERIFICATION
   // Verify the user is ok with the provided inputs
   console.log(chalk.bold.yellow('\nPLEASE VERIFY THE BELOW PARAMETERS\n'));
-  console.log('  Deploying protection market for:   Curve 3Crypto');
-  console.log(`  Deployer address:                  ${signer.address}`);
-  console.log(`  Deploying to network:              ${hre.network.name}`);
-  console.log(`  Gas Price:                         ${JSON.stringify(overrides.gasPrice)}`);
+  console.table({
+    'Deploying protection market for': `${name}`,
+    'Deployer address:              ': `${signer.address}`,
+    'Deploying to network:          ': `${hre.network.name}`,
+    'Underlying token:              ': `${underlyingAddress}`,
+    'Gas price:                     ': `${JSON.stringify(overrides)}`,
+  });
 
   const response = await waitForInput('\nDo you want to continue with deployment? y/N\n');
   if (response !== 'y') {
@@ -66,7 +87,7 @@ async function main(): Promise<void> {
     '0', // multiplierPerYear of 0% = 0 gives 2% borrow rate at kink
     '650000000000000000', // jumpMultiplierPerYear of 65% = 6.5e17 gives 15% borrow rate at 100% utilization
     '800000000000000000', // kink of 0.8 = 8e17 = sets the model kink at 80% utilization
-    '0x1725d89c5cf12F1E9423Dc21FdadC81C491a868b', // Cozy multisig
+    cozyMultiSig, // Cozy multisig
   ];
 
   const irModel: Contract = await irModelFactory.deploy(...constructorArgs);
@@ -75,38 +96,36 @@ async function main(): Promise<void> {
 
   // DEPLOY TRIGGER
   // Get instance of the Trigger ContractFactory with our signer attached
-  const triggerFactory: ContractFactory = await hre.ethers.getContractFactory('CurveThreeTokens', signer);
+  const triggerFactory: ContractFactory = await hre.ethers.getContractFactory('CurveThreeTokenBasePool', signer);
 
   // Deploy the trigger contract (last constructor parameter is specific to the mock trigger contract)
   const triggerParams = [name, symbol, description, platformIds, recipient, curvePoolAddress];
   const trigger: Contract = await triggerFactory.deploy(...triggerParams);
   await trigger.deployed();
-  logSuccess(`CurveThreeTokens trigger deployed to ${trigger.address}`);
+  logSuccess(`CurveThreeTokenBasePool trigger deployed to ${trigger.address}`);
 
   // VERIFY UNDERLYING
-  // Let's choose ETH as the underlying, so first we need to check if there's a ETH Money Market.
   // We know that Money Markets have a trigger address of the zero address, so we use that to query the Comptroller
   // for the Money Market address
-  const ethAddress = getContractAddress('ETH', chainId);
   const comptrollerAddress = getContractAddress('Comptroller', chainId);
   const comptroller = new Contract(comptrollerAddress, comptrollerAbi, signer); // connect signer for sending transactions
-  const cozyEthAddress = await comptroller.getCToken(ethAddress, AddressZero);
+  const cozyMMAddress = await comptroller.getCToken(underlyingAddress, AddressZero);
 
   // If the returned address is the zero address, a money market does not exist and we cannot deploy a protection
   // market with ETH as the underlying
-  if (cozyEthAddress === AddressZero) {
-    logFailure('No ETH Money Market exists. Exiting script');
+  if (cozyMMAddress === AddressZero) {
+    logFailure('Money Market does not exist. Exiting script');
     return;
   }
-  logSuccess(`Safe to continue: Found ETH Money Market at ${cozyEthAddress}`);
+  logSuccess(`Safe to continue: Found Money Market at ${cozyMMAddress}`);
 
   // DEPLOY PROTECTION MARKET
-  // If we're here, a ETH Money Market exists, so it's safe to create our new Protection Market
+  // If we're here, a Money Market exists, so it's safe to create our new Protection Market
   const tx = await comptroller['deployProtectionMarket(address,address,address)'](
-    ethAddress,
+    underlyingAddress,
     trigger.address,
     irModel.address,
-    overrides.gasPrice
+    { ...overrides, gasLimit: '8000000' }
   );
   console.log(`Creating Protection Market in transaction ${tx.hash}`);
