+++ scripts/shared/convex-pools.sh
